backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
websockets==14.2
aiofiles==24.1.0
Pillow==11.1.0
ffmpeg-python==0.2.0
pypandoc==1.14
python-docx==1.1.2
PyPDF2==3.0.1
markdown==3.7
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
